chore(suite): move normalizeVersion from suite pkg to utils pkg
