+++ packages/suite/src/utils/suite/__tests__/build.test.ts
@@ -1,7 +1,4 @@
-/* eslint-disable global-require, @typescript-eslint/no-var-requires */
-import { normalizeVersion } from '../build';
-import { resolveStaticPath } from '@trezor/utils';
-
+/* eslint-disable @typescript-eslint/no-var-requires, global-require */
 const OLD_ENV = { ...process.env };
 
 describe('build', () => {
@@ -42,46 +39,6 @@ describe('build', () => {
             expect(isDev).toEqual(true);
         });
     });
-
-    describe('normalizeVersion', () => {
-        it('removes preceding zeros from versions to normalize it', () => {
-            expect(normalizeVersion('2020.05.13-beta')).toEqual('2020.5.13-beta');
-            expect(normalizeVersion('2022.12.01-beta')).toEqual('2022.12.1-beta');
-            expect(normalizeVersion('3000.04.04-beta')).toEqual('3000.4.4-beta');
-            expect(normalizeVersion('3000.04.04')).toEqual('3000.4.4');
-            expect(normalizeVersion('3000.04.0')).toEqual('3000.4.0');
-        });
-
-        it('does nothing with normalized versions', () => {
-            expect(normalizeVersion('20.11.0')).toEqual('20.11.0');
-            expect(normalizeVersion('20.11.1')).toEqual('20.11.1');
-        });
-    });
-
-    describe('resolve static path', () => {
-        beforeEach(() => {
-            jest.resetModules();
-        });
-
-        afterEach(() => {
-            process.env = { ...OLD_ENV };
-        });
-
-        it('should return static path', () => {
-            process.env.ASSET_PREFIX = '';
-            expect(resolveStaticPath('mypath')).toBe('/static/mypath');
-        });
-
-        it('should return static path even if ASSET_PREFIX is undefined', () => {
-            process.env.ASSET_PREFIX = undefined;
-            expect(resolveStaticPath('mypath')).toBe('/static/mypath');
-            expect(resolveStaticPath('/mypath')).toBe('/static/mypath');
-        });
-
-        it('should return static path prefixed with ASSET_PREFIX', () => {
-            process.env.ASSET_PREFIX = 'brachName';
-            expect(resolveStaticPath('mypath')).toBe('brachName/static/mypath');
-            expect(resolveStaticPath('/mypath')).toBe('brachName/static/mypath');
-        });
-    });
 });
+
+export {};
