+++ packages/utils/src/versionUtils.ts
@@ -62,3 +62,7 @@ export const isEqual = (versionX: VersionInput, versionY: VersionInput) => {
  */
 export const isNewerOrEqual = (versionX: VersionInput, versionY: VersionInput) =>
     isNewer(versionX, versionY) || isEqual(versionX, versionY);
+
+export const normalizeVersion = (version: string) =>
+    // remove any zeros that are not preceded by Latin letters, decimal digits, underscores
+    version.replace(/\b0+(\d)/g, '$1');
