+++ packages/utils/tests/version.test.ts
@@ -1,4 +1,4 @@
-import { isNewer, isNewerOrEqual, isEqual } from '../src/versionUtils';
+import { isNewer, isNewerOrEqual, isEqual, normalizeVersion } from '../src/versionUtils';
 
 describe('Version Utils', () => {
     describe('is newer', () => {
@@ -82,6 +82,21 @@ describe('Version Utils', () => {
         it('it should return true [1, 1, 1], [1, 1, 1]', () => {
             const result = isNewerOrEqual([1, 1, 1], [1, 1, 1]);
             expect(result).toBe(true);
+        });
+    });
+
+    describe('normalizeVersion', () => {
+        it('removes preceding zeros from versions to normalize it', () => {
+            expect(normalizeVersion('2020.05.13-beta')).toEqual('2020.5.13-beta');
+            expect(normalizeVersion('2022.12.01-beta')).toEqual('2022.12.1-beta');
+            expect(normalizeVersion('3000.04.04-beta')).toEqual('3000.4.4-beta');
+            expect(normalizeVersion('3000.04.04')).toEqual('3000.4.4');
+            expect(normalizeVersion('3000.04.0')).toEqual('3000.4.0');
+        });
+
+        it('does nothing with normalized versions', () => {
+            expect(normalizeVersion('20.11.0')).toEqual('20.11.0');
+            expect(normalizeVersion('20.11.1')).toEqual('20.11.1');
         });
     });
 });
